Start adding custom attributes for the dials so I can define the dimensions in XML rather than trying to be clever and compute everything.
This should make it easier to tweak things, though it's also less robust against varying screen sizes, so I'll probably have to add back in some kind of scaling logic.

And change the easing function so that it doesn't overshoot.

+++ app/src/main/java/com/dane/dni/PolarChromeView.java
@@ -14,9 +14,9 @@ import android.view.View;
  */
 public class PolarChromeView extends View {
 
-    private int circleRadius;
-    private int lineRadius;
-    private int lineWidth;
+    private float circleRadius;
+    private float lineRadius;
+    private float lineWidth;
     private Paint paint;
 
     public PolarChromeView(Context context) {
@@ -25,7 +25,7 @@ public class PolarChromeView extends View {
         paint.setColor(Color.rgb(153, 149, 147));
     }
 
-    public void setSizeParams(int circleRadius, int lineRadius, int lineWidth) {
+    public void setSizeParams(float circleRadius, float lineRadius, float lineWidth) {
         this.circleRadius = circleRadius;
         this.lineRadius = lineRadius;
         this.lineWidth = lineWidth;
